% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHCnvex-package.R
\docType{package}
\name{MHCnvex}
\alias{MHCnvex}
\title{MHCnvex}
\section{main.R}{

}

\section{blend.R}{

}

\section{exon.R}{

}

\section{filter.R}{

}

\section{gc.R}{

}

\section{mismatch.R}{

}

\section{models.R}{

}

\section{output.R}{

}

\section{peak.R}{

}

\section{plots.R}{

}

\section{ref.R}{

}

\section{tile.R}{

}

\section{tilebaf.R}{

}

